Update project to use LocalDate instead of Date, which is deprecated.
This simplifies things considerably, as Date is technically Deprecated, and is
actually a DateTime. We only care about the date itself as this is just used
for birthdays. Additionally I have updated the DOB validator to handle null
values and updated the test that returned an invalid date, as well as adding
boundary cases for dates, and a test for invalid dates.