Created fact type enum, wrote filtering query for getting each type of fact
Created a fact type enum to make queries nicer as the discriminator values line up with the ordinal values of the enum. Added a fact_type coloumn to the fact entity to allow for filtering to occur using the JPA queries. Created service and repository and documented them (and a location service that was missing documentation as well).