Created email service method to send the basic become a federation manager email.
created the service method that will send the default templated email with the link (not yet implemented - diff task) to become a federation manger. I also renamed the FederationManagerGenerator.java to FederationManagerInvite.java as it makes more sense. Created user service method to handle the invitation and do the testing to see which email link needs to be sent.