Created front end page for inviting a user to federation manager.
Created the page that displays users to be invited to become federation manager, it uses the existing search + pagination structure. It's not finished in styling or testing but commiting so Angela can branch off to complete linking the invite button to sending email functionality. Also but the endpoint behind fed manager role.